Makefile: set GCC5_AARCH64_PREFIX relative to $WORKSPACE
This way it's sufficient to install gcc in src/tools/gcc (or have gcc
point to whereever the compiler was installed). It is also consistent
with how acpica is installed.

And also do not clean acpi tools upon make clean anymore, as it's long
to build and it's just an external tool.

Signed-off-by: Willy Tarreau <w@1wt.eu>